// models/core.d.ts
export declare class RemotePagerParams {
    root: string;
    page: string;
    total: string;
    records: string;
    repeatitems: boolean;
    cell: string;
    id: string;
}
export declare class ClientPagerParams {
    sidx: string;
    pageSizes: number[];
    pageSize: number;
    sord: string;
    pageStart: number;
}
export declare class DataSourceProperties {
    url: string;
}